Feat does not allow deleting level with developers

+++ app/Http/Controllers/NivelController.php
@@ -66,6 +66,10 @@ class NivelController extends Controller
      */
     public function destroy(Nivel $nivel)
     {
+        if ($nivel->has('desenvolvedores')) {
+            return response()->json(['error' => 'Um (ou mais) desenvolvedor(es) associado a este nivel'], 501);
+        }
+
         $nivel->delete();
 
         return response()->json(null, 204);
